+++ src/Sentry.Unity.Editor/ScriptableSentryUnityOptionsEditor.cs
@@ -19,7 +19,15 @@ namespace Sentry.Unity.Editor
             EditorGUILayout.Toggle("Enable Sentry SDK", options.Enabled);
             EditorGUILayout.TextField("DSN", options.Dsn);
             EditorGUILayout.Toggle("Capture In Editor", options.CaptureInEditor);
+
+            EditorGUILayout.Space();
+            EditorGUI.DrawRect(EditorGUILayout.GetControlRect(false, 1), Color.gray);
+            EditorGUILayout.Space();
+
             EditorGUILayout.FloatField("Traces Sample Rate", (float)options.TracesSampleRate);
+            EditorGUILayout.Toggle("Auto Startup Traces", options.AutoStartupTraces);
+            EditorGUILayout.Toggle("Auto Scene Load Traces", options.AutoSceneLoadTraces);
+            EditorGUILayout.Toggle("Attach Stacktrace", options.AutoAwakeTraces);
 
             EditorGUILayout.Space();
             EditorGUI.DrawRect(EditorGUILayout.GetControlRect(false, 1), Color.gray);
@@ -60,6 +68,14 @@ namespace Sentry.Unity.Editor
             EditorGUI.DrawRect(EditorGUILayout.GetControlRect(false, 1), Color.gray);
             EditorGUILayout.Space();
 
+            EditorGUILayout.LabelField("Application Not Responding", EditorStyles.boldLabel);
+            EditorGUILayout.Toggle("Enable ANR Detection", options.AnrDetectionEnabled);
+            EditorGUILayout.IntField("ANR Timeout [ms]", options.AnrTimeout);
+
+            EditorGUILayout.Space();
+            EditorGUI.DrawRect(EditorGUILayout.GetControlRect(false, 1), Color.gray);
+            EditorGUILayout.Space();
+
             EditorGUILayout.Toggle("iOS Native Support", options.IosNativeSupportEnabled);
             EditorGUILayout.Toggle("Android Native Support", options.AndroidNativeSupportEnabled);
             EditorGUILayout.Toggle("Windows Native Support", options.WindowsNativeSupportEnabled);
@@ -72,10 +88,10 @@ namespace Sentry.Unity.Editor
             EditorGUILayout.Space();
 
             EditorGUILayout.LabelField("Options Configuration", EditorStyles.boldLabel);
-            EditorGUILayout.ObjectField("Runtime Configuration", options.OptionsConfiguration,
+            EditorGUILayout.ObjectField("Runtime Configuration", options.RuntimeOptionsConfiguration,
                 typeof(SentryRuntimeOptionsConfiguration), false);
-            EditorGUILayout.ObjectField("Buildtime Configuration", options.BuildtimeOptionsConfiguration,
-                typeof(SentryBuildtimeOptionsConfiguration), false);
+            EditorGUILayout.ObjectField("Build Time Configuration", options.BuildTimeOptionsConfiguration,
+                typeof(SentryBuildTimeOptionsConfiguration), false);
 
             EditorGUILayout.Space();
             EditorGUI.DrawRect(EditorGUILayout.GetControlRect(false, 1), Color.gray);