+++ src/Sentry.Unity/SentryUnityOptions.cs
@@ -31,6 +31,16 @@ namespace Sentry.Unity
         public bool Enabled { get; set; } = true;
 
         /// <summary>
+        /// "Whether the SDK should automatically create traces during startup."
+        /// </summary>
+        public bool AutoStartupTraces { get; set; } = true;
+
+        /// <summary>
+        /// "Whether the SDK should automatically create traces when loading scenes."
+        /// </summary>
+        public bool AutoSceneLoadTraces { get; set; } = true;
+
+        /// <summary>
         /// Whether Sentry events should be captured while in the Unity Editor.
         /// </summary>
         // Lower entry barrier, likely set to false after initial setup.
@@ -107,6 +117,11 @@ namespace Sentry.Unity
         /// Whether the SDK should automatically add breadcrumbs per LogType
         /// </summary>
         public Dictionary<LogType, bool> AddBreadcrumbsForLogType { get; set; }
+
+        /// <summary>
+        /// The duration in [ms] for how long the game has to be unresponsive before an ANR event is reported.
+        /// </summary>
+        public TimeSpan AnrTimeout { get; set; } = TimeSpan.FromSeconds(5);
 
         /// <summary>
         /// Whether the SDK should add native support for iOS